% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{as.double.ValueWithUncertainty}
\alias{as.double.ValueWithUncertainty}
\title{Coerce to a numeric}
\usage{
\method{as.double}{ValueWithUncertainty}(x, ...)
}
\arguments{
\item{x}{object to be coerced or tested.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
S3 method for \code{ValueWithUncertainty} objects
}
\examples{

}
