% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sample.R
\name{gen_sample}
\alias{gen_sample}
\title{Method for generating samples of uncertain inputs}
\usage{
gen_sample(v, n = 1, fixed = FALSE, ...)
}
\arguments{
\item{fixed}{Disable model and return value, default is false}
}
\value{
Returns a sample of values based on the model
}
\description{
HOLD Generate a sample of inputs base on the model
}
\references{
TBC
}
