% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sample.R
\name{gen_sample.ValueWithUncertainty}
\alias{gen_sample.ValueWithUncertainty}
\title{Method for generating samples of uncertainty inputs}
\usage{
\method{gen_sample}{ValueWithUncertainty}(v, n = 1, fixed = FALSE, ...)
}
\arguments{
\item{fixed}{Disable model and return value, default is false}
}
\value{
HOLD Returns a sample of values with uncertainty based on the model
}
\description{
HOLD Generate a sample of inputs with uncertainty based on the model
}
\references{
TBC
}
