% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R, R/summary.R
\name{Ops.ValueWithUncertainty}
\alias{Ops.ValueWithUncertainty}
\alias{Summary.ValueWithUncertainty}
\title{Ops Value with Uncertainty}
\usage{
\method{Ops}{ValueWithUncertainty}(e1, e2)

\method{Summary}{ValueWithUncertainty}(..., na.rm = FALSE)
}
\arguments{
\item{e1}{Parameter definition tbc}

\item{e2}{Parameter definition tbc}
}
\value{
tbc output from function
}
\description{
Ops

Summary of Value with Uncertainty Export Functions
}
\details{
\subsection{\code{Ops}}{
Boolean operators drop the uncertainty (showing a warning once) and operate on the
numeric values.
Any \code{ValueWithUncertainty} operand is automatically coerced to numeric
with no uncertainty.}


}
\examples{
y <- ValueWithUncertainty(1, 10, 100)
x / sqrt(y) + y * sin(x)

# numeric values automatically coerced this type to numeric and drop
# uncertainty
x^2

# boolean operators drop uncertainty
y > x
c(min(x), max(x))
range(x)
sum(y)
prod(y)
}
\references{
\link{TBC}

\link{TBC}
}
