% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.ValueWithUncertainty}
\alias{summary.ValueWithUncertainty}
\title{TODO:export}
\usage{
\method{summary}{ValueWithUncertainty}(x, ..., digits = max(3L, getOption("digits") - 3L))
}
\description{
TODO:export
}
