% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValueWithUncertainty.R
\name{ValueWithUncertainty}
\alias{ValueWithUncertainty}
\title{ValueWithUncertainty}
\usage{
ValueWithUncertainty(p, digits = 3)
}
\arguments{
\item{p}{A value to report}

\item{digits}{The number of decimal places}
}
\value{
Returns string of the value \code{p}

Returns p with \code{digits} decimal places
}
\description{
\code{ValueWithUncertainty} does something.
}
\examples{

ValueWithUncertainty(1) # returns "1.000"

}
